refactor(web): top jurors components modularization
