+++ web/src/pages/Home/TopJurors/Header/Rewards.tsx
@@ -1,0 +1,28 @@
+import React from "react";
+import styled, { css } from "styled-components";
+import { landscapeStyle } from "styles/landscapeStyle";
+import WithHelpTooltip from "pages/Dashboard/WithHelpTooltip";
+
+const Container = styled.div`
+  ${landscapeStyle(
+    () =>
+      css`
+        width: calc(60px + (240 - 60) * (min(max(100vw, 375px), 1250px) - 375px) / 875);
+      `
+  )}
+`;
+
+const totalRewardsTooltipMsg =
+  "Users have an economic interest in serving as jurors in Kleros: " +
+  "collecting the Juror Rewards in exchange for their work. Each juror who " +
+  "is coherent with the final ruling receive the Juror Rewards composed of " +
+  "arbitration fees (ETH) + PNK redistribution between jurors.";
+
+const Rewards: React.FC = () => (
+  <Container>
+    <WithHelpTooltip place="top" tooltipMsg={totalRewardsTooltipMsg}>
+      <label> Total Rewards </label>
+    </WithHelpTooltip>
+  </Container>
+);
+export default Rewards;
